main.py
from telegram.ext import Updater, CommandHandler, MessageHandler, Filters
from diffusers import StableDiffusionPipeline
from torch import autocast
import telebot
from PIL import Image
from fastapi import FastAPI, Request, Form
from telebot import TeleBot
from telebot.types import InputMediaPhoto

# Load model
pipe = StableDiffusionPipeline.from_pretrained("CompVis/stable-diffusion-v1-4")

# Bot configuration
bot = TeleBot("YOUR_BOT_TOKEN")

# Feature flags
enable_ui = False
enable_multiple_formats = False
enable_image_merge = False
enable_image_edit = False

# Bot commands
def generate_image(update, context):
    prompt = update.message.text
    with autocast("cuda"):
        image = pipe(prompt)["sample"][0]
        if enable_multiple_formats:
            # Add code to generate and send image in different formats
            ...
        else:
            update.message.reply_photo(image)

def handle_message(update, context):
    if enable_ui:
        # Add code to handle user input and navigate through the UI
        ...
    else:
        generate_image(update, context)

# Webhook server
app = FastAPI()

@app.post("/webhook")
async def handle_webhook(request: Request):
    data = await request.json()
    message = data["message"]
    chat_id = message["chat"]["id"]

    # Proses pesan dan generate gambar
    prompt = message["text"]
    with autocast("cuda"):
        image = pipe(prompt)["sample"][0]
        image_bytes = io.BytesIO()
        image.save(image_bytes, format="PNG")

    # Kirim gambar ke Telegram
    bot.send_photo(chat_id, InputMediaPhoto(image_bytes.getvalue()))

    return {"success": True}

# Main loop
if __name__ == "__main__":
    # Bot updater
    updater = Updater(bot.token)
    updater.dispatcher.add_handler(CommandHandler("generate", generate_image))
    updater.dispatcher.add_handler(MessageHandler(Filters.text, handle_message))
    updater.start_polling()

    # Webhook server
    uvicorn.run(app, host="0.0.0.0", port=8181)