requirements.txt
torch
diffusers
telegram
telebot
Pillow